+++ packages/integrations/components/pico/index.js
@@ -1,28 +1,37 @@
-import React, { useState, useEffect, useCallback } from 'react';
+import React, { useEffect, useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
-import Helmet from 'react-helmet';
+import getLogService from '@irvingjs/services/logService';
+import usePollForNode from './usePollForNode';
+import useWidgetScript from './useWidgetScript';
 import PicoObserver from './observer';
 // Pico.
 import {
   actionPicoLoaded,
   actionUpdatePicoPageInfo,
 } from '../../actions/picoActions';
-import { picoLoadedSelector } from '../../selectors/picoSelector';
-// Coral.
 import {
-  actionReceiveCoralLogoutRequest,
-} from '../../actions/coralActions';
-import { tokenSelector } from '../../selectors/coralSelector';
+  picoLoadedSelector,
+  picoScriptAddedSelector,
+} from '../../selectors/picoSelector';
 // Utility functions.
-import mountPicoNodes from './utils';
+import {
+  isPicoMounted,
+  mountPicoNodes,
+} from './utils';
+
+const log = getLogService('irving:pico');
 
 const Pico = (props) => {
   const {
     pageInfo,
     publisherId,
     tiers,
+    widgetUrl,
   } = props;
+
+  // Inject the widget script into the DOM.
+  useWidgetScript(widgetUrl, publisherId);
 
   /**
    * Ensure page info has the right format.
@@ -38,17 +47,21 @@ const Pico = (props) => {
     url: window.location.href,
   };
 
-  // Create a state value that is updated when the `pico-init` event is fired.
-  const [picoInitialized, setPicoInitialized] = useState(false);
   // Create the dispatch function.
   const dispatch = useDispatch();
-  // Create a state value that prevents multiple `UPDATE_PICO_PAGE_INFO` actions
-  // from being fired on a single `LOCATION_CHANGE`.
-  const [picoPageInfoUpdated, setPicoPageInfoUpdated] = useState(false);
   // Create a function that dispatches the current page info for the saga to respond to.
   const dispatchUpdatePicoPageInfo = useCallback(
     (payload) => {
-      setPicoPageInfoUpdated(true);
+      // Trigger the page visit with Pico.
+      const pollForFn = setInterval(() => {
+        if ('function' === typeof window.pico) {
+          // Prevent future polling events.
+          clearInterval(pollForFn);
+          // Trigger the visit.
+          window.pico('visit', payload);
+        }
+      }, 250);
+
       return dispatch(actionUpdatePicoPageInfo(payload));
     },
     [dispatch]
@@ -58,72 +71,76 @@ const Pico = (props) => {
     () => dispatch(actionPicoLoaded()),
     [dispatch]
   );
+
   // Grab the `loaded` value from the `pico` branch of the state tree.
   const picoLoaded = useSelector(picoLoadedSelector);
 
-  // Retrieve the Coral SSO token from the Redux store.
-  const coralToken = useSelector(tokenSelector);
+  // Grab the `scriptAdded` value from the `pico` branch of the state tree.
+  const picoScriptAdded = useSelector(picoScriptAddedSelector);
 
+  // Add listeners for pico events.
   useEffect(() => {
-    // See if the Pico widget container exists in the DOM.
-    const widgetContainer = document.getElementById('pico-widget-container');
-
-    if (widgetContainer) {
-      // Prevent the widget script from being added more than once.
-      setPicoInitialized(true);
-
-      // Ensure the target nodes are only mounted once on the initial server load.
-      if (
-        ! document.getElementById('PicoSignal-container') &&
-        ! document.getElementById('PicoRule-button') &&
-        ! picoLoaded
-      ) {
-        mountPicoNodes();
-        // Trigger the visit.
-        window.pico('visit', picoPageInfo);
-        // Update the `pico` branch of the state tree and set `loaded` to true.
+    // Dispatch an action to update the integrations.pico.loaded state in Redux.
+    const loadHandler = () => {
+      if (! picoLoaded) {
+        log.info('Pico: Running load handler.');
         dispatchPicoLoaded();
       }
-    }
-
-    // Only update the store once Pico has loaded and if it has not already been
-    // updated in this component's lifecycle phase.
-    if (picoLoaded && ! picoPageInfoUpdated) {
-      dispatchUpdatePicoPageInfo(picoPageInfo);
-    }
-
-    const initHandler = () => {
-      setPicoInitialized(true);
     };
 
-    // On component hydration, add an event listener to watch for the script's init event.
-    window.document.addEventListener('pico-init', initHandler);
+    log.info('Pico: Event listeners added.');
+    // The pico.loaded event fires after the Pico object is fully initialized.
+    window.addEventListener('pico.loaded', loadHandler);
 
-    return () => window.document.removeEventListener('pico-init', initHandler);
-  }, [picoInitialized, picoLoaded, picoPageInfoUpdated]);
+    return () => {
+      window.removeEventListener('pico.loaded', loadHandler);
+    };
+  }, []);
 
-  // Load the script for the first time.
-  if (! picoInitialized) {
+  // Mount an effect that triggers the initial visit once `picoScriptAdded` has
+  // been set to true and only if `picoLoaded` has not been set to true yet.
+  useEffect(() => {
+    if (picoScriptAdded) {
+      log.info('Pico: Dispatching page visit.');
+      // Dispatch the initial visit to trigger the `pico.loaded` event.
+      dispatchUpdatePicoPageInfo(picoPageInfo);
+    }
+  }, [picoScriptAdded, picoPageInfo.url]);
+
+  // Poll for the existence of the Pico widget container.
+  const widgetContainer = usePollForNode('#pico-widget-container');
+
+  // Mount an effect that watches for the status of `picoLoaded` and the polled
+  // widget container node in order to determine when to mount the Pico nodes
+  // into the DOM.
+  useEffect(() => {
+    log.info(
+      'Pico: Attempting to mount nodes:',
+      [widgetContainer, isPicoMounted(), picoLoaded]
+    );
+
+    // Ensure the target nodes are only mounted once on the initial server load.
+    if (widgetContainer && ! isPicoMounted()) {
+      log.info('Pico: Mounting nodes.');
+      mountPicoNodes();
+    }
+  }, [picoLoaded, widgetContainer]);
+
+  if (picoScriptAdded) {
+    // Inject the Pico Signal into the DOM.
+    log.info('Pico: Returning observer component.');
+
     return (
-      <Helmet>
-        <script>
-          {`(function(p,i,c,o){var n=new Event("pico-init");i[p]=i[p]||function(){(i[p].queue=i[p].queue||[]).push(arguments)},i.document.addEventListener("pico-init",function(e){var t=i.Pico.getInstance(e,{publisherId:o,picoInit:n},i);t.handleQueueItems(i[p].queue),i[p]=function(){return t.handleQueueItems([arguments])}},!1);var e=i.document.createElement("script"),t=i.document.getElementsByTagName("script")[0];e.async=1,e.defer=1,e.src=c,e.onload=function(e){return i.Pico.getInstance(e,{publisherId:o,picoInit:n},i)},t.parentNode.insertBefore(e,t)})("pico",window,"https://widget.pico.tools/wrapper.min.js", "${publisherId}");`}
-        </script>
-      </Helmet>
+      <PicoObserver tiers={tiers} />
     );
   }
 
-  return (
-    <PicoObserver
-      tiers={tiers}
-      accessToken={coralToken}
-      logoutRequestAction={actionReceiveCoralLogoutRequest}
-    />
-  );
+  return null;
 };
 
 Pico.defaultProps = {
   tiers: [],
+  widgetUrl: 'https://widget.pico.tools',
 };
 
 Pico.propTypes = {
@@ -136,6 +153,7 @@ Pico.propTypes = {
   }).isRequired,
   publisherId: PropTypes.string.isRequired,
   tiers: PropTypes.array,
+  widgetUrl: PropTypes.string,
 };
 
 export default Pico;