+++ packages/styled-components/components/bodyWrapper/index.js
@@ -1,18 +1,14 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { Helmet } from 'react-helmet';
-import classNames from 'classnames';
 import Loader from 'components/loader';
 import useStandardProps from '@irvingjs/styled/hooks/useStandardProps';
 import {
   standardPropTypes,
-  standardDefaultProps,
+  getStandardDefaultProps,
 } from '@irvingjs/styled/types/propTypes';
 import * as defaultStyles from './themes/default';
 
 const BodyWrapper = (props) => {
   const {
-    bodyClasses,
     children,
     theme,
   } = props;
@@ -21,12 +17,9 @@ const BodyWrapper = (props) => {
 
   return (
     <Loader>
-      <Helmet>
-        <body className={classNames(bodyClasses)} />
-      </Helmet>
       <Main
         {...standardProps}
-        role="main"
+        id="content"
       >
         {children}
       </Main>
@@ -36,18 +29,10 @@ const BodyWrapper = (props) => {
 
 BodyWrapper.propTypes = {
   ...standardPropTypes,
-  /**
-   * Additional classes to apply to the <body> tag using react-helmet.
-   */
-  bodyClasses: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.string),
-    PropTypes.string,
-  ]),
 };
 
 BodyWrapper.defaultProps = {
-  ...standardDefaultProps,
-  bodyClasses: [],
+  ...getStandardDefaultProps(),
   theme: defaultStyles,
 };