+++ packages/styled/components/siteThemeProvider/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { ThemeProvider } from 'styled-components';
+import { recursivelyBuildObjectTree } from '../../utils/siteTheme';
 
 /**
  * Component for using the `styled-components` package's ThemeProvider.
@@ -12,26 +13,21 @@ const SiteThemeProvider = (props) => {
   } = props;
 
   return (
-    <ThemeProvider theme={theme}>
+    <ThemeProvider theme={recursivelyBuildObjectTree(theme, theme)}>
       {children}
     </ThemeProvider>
   );
-};
-
-SiteThemeProvider.defaultProps = {
-  children: [],
-  theme: {},
 };
 
 SiteThemeProvider.propTypes = {
   /**
    * Children of the component.
    */
-  children: PropTypes.array,
+  children: PropTypes.array.isRequired,
   /**
-   * Site theme.
+   * Children of the component.
    */
-  theme: PropTypes.object,
+  theme: PropTypes.object.isRequired,
 };
 
 export default SiteThemeProvider;