+++ packages/wp-admin-bar/index.js
@@ -1,3 +1,0 @@
-import AdminBar from './components/adminBar';
-
-export default AdminBar;