+++ packages/wp-admin-bar/jest.config.js
@@ -1,7 +1,0 @@
-const baseConfig = require('../../jest.config.base.js');
-
-module.exports = {
-  ...baseConfig,
-  name: '@irvingjs/wp-admin-bar',
-  displayName: '@irvingjs/wp-admin-bar',
-};